focus_guard_server/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
aiofiles
httpx
google-generativeai
python-dotenv
opencv-python
Pillow
aiosqlite
numpy
torch
torchvision
mediapipe
